refactor: participate_ctn_refactor완료
